UI fixes

+++ RoomReservationSystem/Reservation/Reservation.cs
@@ -12,7 +12,7 @@ namespace Reservations
         public DateTime reservationDate { set; get; }
         public int clientId { set; get; }
         public int roomId { set; get; }
-        public float price { set; get; }
+        public double price { set; get; }
         public DateTime checkInDate { set; get; }
         public DateTime checkOutDate { set; get; }
